pyproject.toml
[project]
name = "one"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "django",
  "django-auto-prefetch",
  "django-cleanup",
  "django-extensions",
  "django-rosetta",
  "django-o365mail",
  "django-money[exchange]",
  "django-upgrade",
  "django-markdownify",
  "django_htmx",
  "boto3",
  "psycopg",
  "pillow",
  "huey",
  "redis",
  "hiredis",
  "geoip2",
  "Werkzeug",
  "django-browser-reload",
  "docutils",
  "django-debug-toolbar",
  "deepl",
  "Jinja2",
  "qrcode[pil]",
  "Pygments",
  "python-barcode[images]",
  "django-taggit",
  "gunicorn",
  "ipython",
  "coverage[toml]",
  "factory-boy",
  "black",
  "pyclean",
  "isort",
  "ruff",
  "djade",
  "djlint",
  "ipdb",
  "django-stubs[compatible-mypy]",
  "environs",
  "django-allauth[socialaccount]",
  "channels-redis",
  "channels[daphne]",
  "pdf2image",
  "django-dbbackup",
  "etsyv3",
  "django-modeltranslation>=0.19.11",
  "django-cors-headers>=4.6.0",
  "pydotplus>=2.0.2",
  "twisted[http2,tls]>=24.11.0",
  "django-storages>=1.14.4",
  "whitenoise[brotli]>=6.8.2",
  "django-minify-html>=1.11.0",
  "stripe>=11.4.1",
  "unittest-parametrize>=1.5.0",
  "django-nested-inline>=0.4.6",
]


[tool.ruff]
exclude = [".venv", "venv", "migrations", "submodules"]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "UP", "B", "SIM", "I"]

[tool.coverage.run]
branch = true
omit = ["manage.py"]

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "one.settings"
